+++ CHANGELOG.md
@@ -10,6 +10,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Added `as` parameter for `stFrom` function
 
+### Fixed
+
+- Fixed builder macro docblocks to use FQDN for an issue within laravel-ide-helper
+
 ## [1.4.0](https://github.com/clickbar/laravel-magellan/tree/1.4.0) - 2023-07-02
 
 ### Improved