+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - Added `as` parameter for `stFrom` function
+- Aligned `stHaving` with `stWhere` to also accept string column and ST expressions, boxes and geometries as value
 
 ### Fixed
 - Fixed builder macro docblocks to use FQDN for an issue within laravel-ide-helper